feat(user): implementar PATCH /api/users/:userId e corrigir fluxo de update
- Controller
  - Lê o parâmetro correto da rota (`params.userId`) e valida UUID
  - Valida campos permitidos (`first_name`, `last_name`, `email`, `password`)
  - Rejeita senha < 6 e e-mail inválido (validator)
  - Retorna `200 OK` com o usuário atualizado via `ok(...)`
  - Usa helpers padronizados (`badRequest`, `ok`, `serverError`)
- Use Case
  - `UpdateUserUseCase.execute(userId, updateUserParams)` com dois parâmetros
  - Normaliza e-mail para minúsculas e, se necessário, prepara hash de senha
  - (Compatível com verificação de e-mail em uso via repositório)
- Repository (PostgresUpdateUserRepository)
  - Monta SQL dinâmica com placeholders corretos
  - **Fix:** `WHERE id = $<len+1>` (antes usava o índice errado)
  - Usa `RETURNING` e devolve **a linha** atualizada (não `rowCount`)
- Infra/ESM & imports
  - Corrigidos caminhos relativos (ex.: `../../db/postgres/helper.js`)
  - Mantida extensão `.js` em todos os imports ESM
- Comportamentos
  - 400: ID ausente/UUID inválido, campos não permitidos, payload vazio, validações
  - 200: usuário atualizado (sem vazar password)
  - 500: apenas para erros inesperados (fallback do `serverError`)
