+++ index.js
@@ -2,6 +2,7 @@ import 'dotenv/config';
 import express from 'express';
 import { CreateUserController } from './src/controllers/create-user.js';
 import { GetUserByIdController } from './src/controllers/get-user-by-id.js';
+import { UpdateUserController } from './src/controllers/update-user.js';
 
 const app = express();
 app.use(express.json());
@@ -11,11 +12,11 @@ app.post('/api/users', async (request, response) => {
   const {statusCode, body} = await createUserController.execute(request)
   response.status(statusCode).send(body)
 })
-app.patch('/api/users/:userId', async (request, response) => {
-  const updateUserController = new UpdateUserController()
-  const {statusCode, body} = await updateUserController.execute(request)
-  response.status(statusCode).send(body)
-})
+app.patch('/api/users/:userId', async (req, res) => { // << :userId
+  const controller = new UpdateUserController();
+  const { statusCode, body } = await controller.execute(req);
+  res.status(statusCode).send(body);
+});
 app.get('/api/users/:userId', async (request, response) => {
   const getUserByIdController = new GetUserByIdController();
   const { statusCode, body } = await getUserByIdController.execute(request);
