--- src/repositories/update-user.js
@@ -1,27 +1,0 @@
-import {PostgresHelper} from '../..db/postgres/helper.js'
-
-export class PostgresUpdateUserRepository {
-    async execute(userId, updateUserParams){
-        const updateFields = []
-        const updateValues = []
-
-        Object.keys(updateUserParams).forEach((key, index) => {
-            updateFields.push(`"${key}" = $${updateValues.length + 1}`)
-            updateValues.push(updateUserParams[key])
-        })
-        updateValues.push(userId)
-
-        const updateQuery = `
-            UPDATE users
-            SET ${updateFields.join(", ")}
-            WHERE id = $${updateFields.length}
-            RETURNING *
-        `
-
-        const updatedUser = await PostgresHelper.query(
-            updateQuery,
-            updateValues
-        )
-        return updatedUser[0]
-    }
-}
